Constant width timestamps
Shows a fixed number of digits in the subsecond fraction in timestamps.

This reverts to using `format_description!`, and will always show UTC
time with an offset of `Z` and local time with an offset of `+/-HH:MM`.
This should still comply with RFC 3339, which specifically calls out
number of digits when talking about ordering timestamp strings.

Closes #55